Trying to improve spell-checking tools.
Human verification of the list of words to be ignored, and splitting of
the files. The listing one is dangerous since there are so many labels...
Being able to ignore all what is related to the listing environment (or
to the coq custom one we use) would probably be better. But I do not know
how to do this.